update: move functions to other files to make deploy.so neat
